rebase:feat/login onto develop

+++ app/src/main/java/com/example/teacherforboss/presentation/ui/auth/signup/api/EmailCheckRequest.kt
@@ -2,9 +2,9 @@ package com.example.teacherforboss.presentation.ui.auth.signup.api
 
 import com.google.gson.annotations.SerializedName
 
-data class EmailCheckRequest (
+data class EmailCheckRequest(
     @SerializedName("emailAuthId")
-    var emailAuthId:Int,
+    var emailAuthId: Long,
     @SerializedName("emailAuthCode")
     var emailAuthCode:String
 )
